% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_calls.R
\name{call_log_ct}
\alias{call_log_ct}
\title{Calls an API, logs the calls, warnings, and errors.}
\usage{
call_log_ct(
  grid_df,
  traceback = FALSE,
  autolog = TRUE,
  show_notes = TRUE,
  compact = FALSE,
  progress_bar = TRUE,
  return_results = TRUE
)
}
\arguments{
\item{grid_df}{A data frame containing the parameters for the API calls.}

\item{traceback}{Logical indicating whether traceback information should be included in the log (default: FALSE).}

\item{autolog}{Logical indicating whether autolog information should be included in the log (default: TRUE).}

\item{show_notes}{Logical indicating whether notes should be included in the log (default: TRUE).}

\item{compact}{Logical indicating whether the log should be compacted (default: FALSE).}

\item{progress_bar}{Logical indicating whether a progress bar should be displayed during processing (default: TRUE).}

\item{return_results}{Logical indicating whether the results of the API calls should be returned (default: TRUE).}
}
\value{
A list containing the results of the API calls for each row of the data frame.
}
\description{
This function calls an API for each row of a given data frame and logs the calls, warnings, and errors.
}
