% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_calls.R
\name{call_log_yt}
\alias{call_log_yt}
\title{Call YouTube API for Multiple Accounts and Store Results}
\usage{
call_log_yt(
  grid_df,
  traceback = FALSE,
  autolog = TRUE,
  show_notes = TRUE,
  compact = FALSE,
  progress_bar = TRUE,
  return_results = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{grid_df}{A dataframe containing the grid of parameters for each account, including 'data' for the data directory,
'accounts' for the account IDs, 'start' for the start date, and 'end' for the end date.}

\item{traceback}{Logical. If TRUE, includes a traceback in error messages. Defaults to FALSE.}

\item{autolog}{Logical. If TRUE, automatically logs messages to a file. Defaults to TRUE.}

\item{show_notes}{Logical. If TRUE, shows notes in the log file. Defaults to TRUE.}

\item{compact}{Logical. If TRUE, uses compact formatting for log messages. Defaults to FALSE.}

\item{progress_bar}{Logical. If TRUE, displays a progress bar. Defaults to TRUE.}

\item{return_results}{Logical. If TRUE, returns the results. Defaults to TRUE.}

\item{verbose}{Logical. If TRUE, displays verbose messages. Defaults to TRUE.}
}
\value{
A list containing the results for each account, or NULL if an error occurs.
}
\description{
This function calls the YouTube API for multiple accounts specified in a grid dataframe and stores the results in JSON format.
}
\details{
This function iterates over each row of the grid dataframe and calls the \code{get_channel_vids_ERRORS} function from the 'tuber' package
to retrieve video details for each account within the specified date range. Results are stored in JSON format in the data directory specified
in the grid dataframe. Error messages and warnings are logged to a file.
}
\examples{
\dontrun{
# Create a grid dataframe
grid_df <- data.frame(
  data = "data_directory",
  accounts = c("account1", "account2"),
  start = "2022-01-01",
  end = "2022-01-31"
)

# Call the YouTube API for multiple accounts
call_log_yt(grid_df)
}

}
