% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_calls.R
\name{call_log_yt_comments}
\alias{call_log_yt_comments}
\title{Call YouTube Comment API for Multiple Video IDs with Logging}
\usage{
call_log_yt_comments(
  video_ids,
  data_dir = NULL,
  auth = "key",
  simplify = FALSE,
  max_n = Inf,
  verbose = TRUE,
  progress_bar = TRUE,
  return_results = TRUE
)
}
\arguments{
\item{video_ids}{A character vector containing YouTube video IDs for which comments need to be retrieved.}

\item{data_dir}{Directory path where log files and fetched comments will be saved.}

\item{auth}{A character string specifying the authentication method. Default is "key".}

\item{simplify}{A logical value indicating whether to simplify the resulting data structure. Default is FALSE.}

\item{max_n}{An integer specifying the maximum number of comments to fetch per video. Default is Inf.}

\item{verbose}{A logical value indicating whether to print detailed progress messages. Default is TRUE.}

\item{progress_bar}{A logical value indicating whether to display a progress bar. Default is TRUE.}

\item{return_results}{A logical value indicating whether to return the fetched comments as a list. Default is TRUE.}
}
\value{
If \code{return_results = TRUE}, a list containing the fetched comments for each video ID; otherwise, NULL.
}
\description{
This function calls the YouTube Comment API for multiple video IDs, retrieves comments, and logs the process.
}
\details{
This function iterates over the provided video IDs, calls the YouTube Comment API for each video,
retrieves comments, and saves the fetched comments to JSON files.
It logs the process, including progress messages, warnings, and errors. If any error occurs during the process,
it logs the error and returns NULL.
}
\seealso{
\code{\link{get_comments_yt}}
}
