% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comments_call_grid.R
\name{comments_call_deduplicate}
\alias{comments_call_deduplicate}
\title{Check If Comments Were Already Collected}
\usage{
comments_call_deduplicate(dt)
}
\arguments{
\item{dt}{A \code{data.table} containing an \code{item_id} column and a \code{save_dir} column. The first element of the \code{save_dir} column is used to determine the directory to check for collected items.}
}
\value{
A \code{data.table} with rows removed where the \code{item_id} is found in the filenames in the specified directory.
}
\description{
This function checks whether items in a given data.table (\code{dt}) have already been collected by comparing \code{item_id} values with those found in comment filenames in a specified directory.
}
\details{
The function retrieves the \code{save_dir} from the \code{dt}, uses a safe version of \code{parse_comment_filenames} to parse filenames in the directory, and removes rows from \code{dt} where the \code{item_id} is found in the parsed filenames.
}
