% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dirs_if}
\alias{create_dirs_if}
\title{Create Directories if They Do Not Exist}
\usage{
create_dirs_if(dirs = NULL)
}
\arguments{
\item{dirs}{A character vector of directory paths to create. Default is \code{NULL}.}
}
\value{
This function does not return any value. It is called for its side effect of creating directories.
}
\description{
This function takes a vector of directory paths and creates those directories if they do not already exist.
}
\details{
The function ensures that all provided directory paths are unique before attempting to create them.
It uses \code{purrr::walk} to iterate over each directory path and create the directory if it does not exist.
If a directory already exists, it is skipped.
}
