% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_call_grid.R
\name{create_filename}
\alias{create_filename}
\title{Create Filenames for Social Media Data Collection}
\usage{
create_filename(
  df,
  platform = c("fb", "ig", "tt", "yt", "tg", "bc", "bs"),
  country_var = NULL,
  party_var = NULL,
  name_var = NULL,
  filename_var = NULL,
  name_sep = "-",
  lowercase = T,
  replace_non_ascii = T,
  filename_sep = "_"
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{platform}{A character vector specifying the platform(s) from which the data is collected. Default is c("fb", "ig", "tt", "yt", "tg", "bc", "bs").}

\item{country_var}{A character string specifying the column name in the data frame that contains country information. Default is NULL.}

\item{party_var}{A character string specifying the column name in the data frame that contains party information. Default is NULL.}

\item{name_var}{A character string specifying the column name in the data frame that contains name information. Default is NULL.}

\item{filename_var}{A character string specifying the column name to store the constructed filenames. Default is NULL.}

\item{name_sep}{A character string specifying the separator for name variables. Default is "-".}

\item{lowercase}{Logical; if TRUE, filenames will be converted to lowercase. Default is TRUE.}

\item{replace_non_ascii}{Logical; if TRUE, non-ASCII characters will be replaced in filenames. Default is TRUE.}

\item{filename_sep}{A character string specifying the separator for filename construction. Default is "_".}
}
\value{
The input data frame \code{df} with an additional column containing constructed filenames.
}
\description{
This function constructs filenames for social media data collection based on specified parameters.
}
\examples{
df <- data.frame(
  country = c("USA", "UK", "Germany"),
  party = c("Democrat", "Republican", "Green"),
  name = c("Joe Biden", "Donald Trump", "Angela Merkel")
)

create_filename(df,
  platform = "fb",
  country_var = "country",
  party_var = "party",
  name_var = "name",
  filename_var = "filename"
)

}
