% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_call_grid.R
\name{create_update_grid}
\alias{create_update_grid}
\title{Create Update Grid}
\usage{
create_update_grid(
  df = df,
  platform = c("fb", "ig", "tt", "yt", "tg", "bc", "bs"),
  country_var = NULL,
  party_var = NULL,
  name_var = NULL,
  filename_var = NULL,
  account_var = NULL,
  start_date = NULL,
  end_date = NULL,
  name_sep = "-",
  lowercase = TRUE,
  replace_non_ascii = TRUE,
  filename_sep = "_",
  sortBy = "date",
  parse = TRUE,
  data_path = NULL,
  count = Inf,
  drop_existing = FALSE,
  cleanup = FALSE
)
}
\arguments{
\item{df}{The data frame containing relevant information for creating the update grid.}

\item{platform}{A character vector specifying the platforms to include in the update grid.
Valid values are "fb" (Facebook), "ig" (Instagram), "tt" (TikTok), "yt" (YouTube), "tg" (Telegram), "bc" (BitChute), and "bs" (BitScope).}

\item{country_var}{The variable in the data frame representing country information.}

\item{party_var}{The variable in the data frame representing party information.}

\item{name_var}{The variable in the data frame representing name information.}

\item{filename_var}{The variable in the data frame representing filename information.}

\item{account_var}{The variable in the data frame representing handle information.
IMPORTANT: For Crowdtangle this is assumed to be the handle, while for all other platforms, a account_id is expected.}

\item{start_date}{The start date for the update grid.}

\item{end_date}{The end date for the update grid.}

\item{name_sep}{The separator used in constructing filenames.}

\item{lowercase}{Logical indicating whether filenames should be converted to lowercase.}

\item{replace_non_ascii}{Logical indicating whether non-ASCII characters should be replaced in filenames.}

\item{filename_sep}{The separator used in filenames.}

\item{sortBy}{The variable used for sorting the update grid.}

\item{parse}{Logical indicating whether to parse the update grid.}

\item{data_path}{The path to the data.}

\item{count}{The count for the update grid.}

\item{drop_existing}{Logical indicating whether to drop existing files.}

\item{cleanup}{Logical. Should the directory \code{path} checked for corrupt jsons and error jsons before parsing? Slows up process.}
}
\value{
A data frame containing the generated update grid.
}
\description{
This function generates an update grid for various platforms based on provided parameters.
}
\details{
This function generates an update grid for specified platforms using the provided parameters.
It first finds the latest stored files, renames variables, merges timeframes, adds time information to filenames,
and creates the update grid based on the specified parameters.
}
