% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_utils.R
\name{ct_create_list_file}
\alias{ct_create_list_file}
\title{Create Crowdtangle List File from Account Dataframe}
\usage{
ct_create_list_file(
  account_df,
  account_var = NULL,
  drop_private = TRUE,
  list_name_var = NULL,
  list_name = NULL,
  platform = c("fb", "ig"),
  file = NULL
)
}
\arguments{
\item{account_df}{A dataframe containing account information.}

\item{account_var}{The name of the column containing the handles.}

\item{drop_private}{Logical, indicating whether to drop rows containing "private" or "profile" in the doubts column. Default is TRUE.}

\item{list_name_var}{Variable indicating the list names.}

\item{list_name}{The name to be assigned to the "List" column.}

\item{platform}{Platform name. Can be "fb" for Facebook or "ig" for Instagram.}

\item{file}{Path to save the resulting dataframe as a CSV file.}
}
\value{
Subsetted dataframe with renamed columns.
}
\description{
This function creates a list file from a dataframe containing account information
about Facebook or Instagram handles, in the format expected by Crowdtangle batch upload
to create a list.
}
