% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_utils.R
\name{ct_getlistaccounts}
\alias{ct_getlistaccounts}
\title{Get accounts from a Crowdtangle list}
\usage{
ct_getlistaccounts(list = NULL, token = NULL, count = Inf)
}
\arguments{
\item{list}{A character string specifying the ID of the Crowdtangle list.}

\item{token}{A character string specifying the Crowdtangle API token. If not provided,
the token will be retrieved from the environment variable "CROWDTANGLE_TOKEN".}

\item{count}{An integer specifying the maximum number of accounts to retrieve. Default is Inf.}
}
\value{
A dataframe containing the retrieved accounts.
}
\description{
This function retrieves accounts from a Crowdtangle list using the Crowdtangle API.
}
