% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_utils.R
\name{ct_merge_list_id}
\alias{ct_merge_list_id}
\title{Merge Crowdtangle list IDs with Account Dataframe}
\usage{
ct_merge_list_id(acc_df, token, handle_var = NULL)
}
\arguments{
\item{acc_df}{A dataframe containing account information.}

\item{token}{A character string specifying the Crowdtangle API token.}

\item{handle_var}{The name of the variable containing account handles in the 'acc_df' dataframe.}
}
\value{
A dataframe with Crowdtangle list IDs merged with the original account dataframe.
}
\description{
This function merges Crowdtangle list IDs with an account dataframe based on account handles.
}
