% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_call_grid.R
\name{drop_existing}
\alias{drop_existing}
\title{Checks for already downloaded JSON files and drops corresponding calls from the grid.}
\usage{
drop_existing(path, grid, filename_var = "filename", recursive = FALSE)
}
\arguments{
\item{path}{The path where the JSON files are stored.}

\item{grid}{The data frame containing the calls to be filtered.}

\item{filename_var}{The variable name in the grid that contains the filenames (default: "filename").}

\item{recursive}{Logical indicating whether to search for files recursively in subdirectories (default: FALSE).}
}
\value{
Returns a modified grid data frame with calls corresponding to existing JSON files dropped.
}
\description{
This function checks for already downloaded JSON files in a specified path based on filename patterns.
It then updates the input grid to drop calls that correspond to existing JSON files.
}
