% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_stored_data.R
\name{find_latest}
\alias{find_latest}
\title{Find Latest Files for Each Account}
\usage{
find_latest(path)
}
\arguments{
\item{path}{The directory path where the files are located.}
}
\value{
A data table containing information about the latest files for each account, including:
\itemize{
\item \code{account_handle}: The handle of the account.
\item \code{account_name}: The name of the account.
\item \code{account_pageAdminTopCountry}: The top country associated with the account.
\item \code{date}: The date of the latest file.
}
}
\description{
This function parses the latest files for each account within a specified directory path.
It returns a data table containing information about the latest files for each account,
including the file name, account details, and the date of the latest file.
}
\details{
This function first checks if the provided directory path exists. If it doesn't exist,
the function stops and displays an error message. It then parses the latest files for each account,
drops unwanted columns, unnests remaining columns, converts the date column to POSIXct if it's not
already in that format, orders the data table by date, and finally selects the row with the maximum
date for each account handle.
}
