% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber_utils.R
\name{get_channel_vids}
\alias{get_channel_vids}
\title{Get Video Details from YouTube Channel}
\usage{
get_channel_vids(
  channel_id = channel_id,
  start_date = NULL,
  end_date = NULL,
  max_results = Inf,
  part = c("contentDetails", "id", "liveStreamingDetails", "localizations", "player",
    "recordingDetails", "snippet", "statistics", "status", "topicDetails"),
  ...
)
}
\arguments{
\item{channel_id}{The ID of the YouTube channel.}

\item{start_date}{The start date to filter videos. Defaults to NULL (no filtering).}

\item{end_date}{The end date to filter videos. Defaults to NULL (no filtering).}

\item{max_results}{Maximum number of results to retrieve. Defaults to Inf (retrieve all results).}

\item{part}{A character vector specifying the parts of the resource to retrieve.
Defaults to c("contentDetails", "id", "liveStreamingDetails", "localizations",
"player", "recordingDetails", "snippet", "statistics", "status", "topicDetails").}

\item{...}{Additional arguments to be passed to underlying functions.}
}
\value{
A list containing details of videos from the YouTube channel.
}
\description{
This function retrieves video details from a YouTube channel based on provided parameters.
}
\examples{
\dontrun{
get_channel_vids(channel_id = "your_channel_id", start_date = "2023-01-01", end_date = "2023-12-31")
}

}
