% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuber_utils.R
\name{get_comments_yt}
\alias{get_comments_yt}
\title{Fetch YouTube Comments for a Video}
\usage{
get_comments_yt(
  video_id = NULL,
  simplify = TRUE,
  max_n = Inf,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{video_id}{A character string specifying the YouTube video ID.}

\item{simplify}{A logical value indicating whether to simplify the results into a data table. Default is \code{TRUE}.}

\item{max_n}{An integer specifying the maximum number of comments to retrieve. Default is \code{Inf}.}

\item{verbose}{A logical value indicating whether to print messages about the fetching progress. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to the \code{tuber:::tuber_GET} function.}
}
\value{
A list of comments or a data table if \code{simplify = TRUE}. Returns an error message if an error occurs.
}
\description{
This function fetches comments for a specified YouTube video using the YouTube Data API.
It retrieves comments in multiple pages and stops fetching when the specified maximum number
of comments is reached. If an error occurs during the API call, the function stops and returns an error message.
}
\examples{
\dontrun{
comments <- get_comments_yt(video_id = "your_video_id", max_n = 100)
}
}
