% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_calls.R
\name{map_yt_comments}
\alias{map_yt_comments}
\title{Map YouTube Comments Collection}
\usage{
map_yt_comments(slices, return_results = TRUE)
}
\arguments{
\item{slices}{A list of data.tables. Each data.table should contain columns \code{item_id} and \code{save_dir}. \code{item_id} represents the identifier for the post or video, and \code{save_dir} represents the directory where the comments JSON file should be saved.}

\item{return_results}{A logical value indicating whether to return the results of the comments
collection. Default is \code{TRUE}.}
}
\value{
If \code{return_results} is \code{TRUE}, a list of results from the comments collection is returned.
If \code{return_results} is \code{FALSE}, the function does not return any value.
}
\description{
This function processes a list of data.tables, each containing information about social media posts
or videos posted within different timeframes. It checks if the \code{save_dir} specified in those
data.tables already contains a JSON file for each post (by \code{item_id}). If not, it collects all
comments for that video or post.
}
\details{
The function first checks each data.table in \code{slices} to determine if comments for the
posts/videos have already been collected by looking for existing JSON files in the
specified \code{save_dir}. It then calls the \code{call_log_yt_comments} function for each data.table to
collect comments for posts/videos that have not been collected yet.
}
