% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_stored_data.R
\name{parse_comment_filenames}
\alias{parse_comment_filenames}
\title{Parse Comment Filenames}
\usage{
parse_comment_filenames(path = NULL, recursive = FALSE)
}
\arguments{
\item{path}{A character string specifying the directory path where JSON files are located. Default is \code{NULL}.}

\item{recursive}{A logical value indicating whether to search for files recursively within subdirectories. Default is \code{FALSE}.}
}
\value{
A \code{tibble} containing the following columns:
\describe{
\item{item_id}{The extracted item ID from the filename.}
\item{plat}{The platform identifier extracted from the filename (e.g., \code{fb}, \code{ig}, \code{tt}, \code{yt}, \code{tg}, \code{bc}, \code{bs}).}
\item{is_comment}{Logical. Indicating that file contains comments.}
\item{full_filepath}{The full file path of each JSON file.}
\item{filenames}{The base filename of each JSON file.}
\item{dirname}{The name of the folder in which the JSONs are stored.}
}
}
\description{
This function parses filenames of JSON files containing comments from various platforms.
It extracts metadata such as platform identifier, download date, download time, and item ID.
}
\details{
The function checks if the provided directory path exists and then lists all JSON files in the directory (and subdirectories if \code{recursive} is \code{TRUE}). It extracts the platform identifier from the beginning of the filename, cleans up the filename to extract the download date and time, and constructs a download datetime. It also extracts the item ID from the cleaned filename.
}
