% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_stored_data.R
\name{parse_data}
\alias{parse_data}
\title{Parse data from specified directory}
\usage{
parse_data(dir = NULL, filepaths = NULL, cleanup = FALSE)
}
\arguments{
\item{dir}{A character string specifying the directory from which data needs to be parsed. Default is NULL.}

\item{filepaths}{A character vector specifying the filepaths to be parsed. Default is NULL.}

\item{cleanup}{Logical. Should the directory \code{path} checked for corrupt jsons and error jsons before parsing? Slows up process.}
}
\value{
A data.table containing parsed data from the specified directory or filepaths.
}
\description{
This function parses data from the specified directory. It first checks if the directory exists. If not, it throws an error.
Then, it removes invalid JSON files from the directory, if any. Next, it parses the filenames and extracts information
from them. Depending on the platform (Facebook, Instagram, or YouTube), it parses the data accordingly using the
appropriate methods.
}
