% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_stored_data.R
\name{parse_filename_strings}
\alias{parse_filename_strings}
\title{Parse Filename Strings to Extract Metadata}
\usage{
parse_filename_strings(filepaths)
}
\arguments{
\item{filepaths}{A character vector of file paths or filenames to be parsed.}
}
\value{
A tibble with the following columns:
\describe{
\item{\code{plat}}{A character vector representing the platform (e.g., "fb", "ig", "tt", "yt", "tg", "bc", "bs").}
\item{\code{country}}{A character vector representing the country code extracted from the filename.}
\item{\code{person}}{A character vector representing the country_party_account extracted from the filename.}
\item{\code{party}}{A character vector representing the party extracted from the filename.}
\item{\code{account_owner}}{A character vector representing the account extracted from the filename.}
\item{\code{from_date}}{A Date vector representing the start date extracted from the filename.}
\item{\code{from_datetime}}{A character vector representing the start datetime in ISO8601 format.}
\item{\code{to_date}}{A Date vector representing the end date extracted from the filename.}
\item{\code{to_datetime}}{A character vector representing the end datetime in ISO8601 format.}
\item{\code{dl_date}}{A Date vector representing the download date extracted from the filename.}
\item{\code{dl_datetime}}{A character vector representing the download datetime in ISO8601 format.}
\item{\code{basenames}}{A character vector of the base names of the filenames.}
\item{\code{full_filepath}}{A character vector of the full file paths provided as input.}
}
}
\description{
This function parses a vector of filenames to extract various metadata components such as platform, dates, times, country, and person/party information. The metadata is returned in a tidy data frame.
}
