% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_stored_data.R
\name{parse_filenames}
\alias{parse_filenames}
\title{Parse filenames of JSON files containing social media data}
\usage{
parse_filenames(path, recursive = FALSE)
}
\arguments{
\item{path}{A character string specifying the directory path where the JSON files are located.}

\item{recursive}{Logical indicating whether to search for files recursively in subdirectories. Default is \code{FALSE}.}
}
\value{
A data frame containing parsed information from the filenames of the JSON files.
}
\description{
This function parses information stored in the filenames of JSON files containing collected social media data.
}
\examples{
parse_filenames(path = ".")


}
