% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_invalid_jsons}
\alias{remove_invalid_jsons}
\title{Remove Invalid JSON Files}
\usage{
remove_invalid_jsons(dir = NULL, filepaths = NULL)
}
\arguments{
\item{dir}{(character) Path to the directory containing JSON files.}

\item{filepaths}{(character) Vector of file paths to specific JSON files. Default is NULL.}
}
\description{
This function removes invalid JSON files by either moving them to a "corrupted_files" sub-directory
or raising an error if no JSON files are found or if the directory does not exist.
}
\details{
The function checks if either a directory containing JSON files ('dir') or direct file paths to JSON files ('filepaths') are provided. It then proceeds to validate the JSON files using RcppSimdJson::validateJSON. Invalid JSON files are moved to a "corrupted_files" sub-directory created in the same directory as the original file.
}
