% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comments_call_grid.R
\name{slice_post_timeframes}
\alias{slice_post_timeframes}
\title{Read JSON post files and split data into timeframes}
\usage{
slice_post_timeframes(
  dir = NULL,
  filepaths = NULL,
  cutoff = NULL,
  recursive = FALSE,
  verbose = TRUE,
  include_latest = TRUE,
  inclued_oldest = TRUE,
  input_tz = "CEST"
)
}
\arguments{
\item{dir}{A character string specifying the directory from which data needs to be parsed.}

\item{filepaths}{A character vector of file paths to JSON files. If specified, overrides the \code{dir} parameter.}

\item{cutoff}{A numeric vector of cutoff times in hours for categorizing the time difference to now.}

\item{recursive}{A logical value indicating whether to search the directory recursively. Default is FALSE.}

\item{verbose}{A logical value indicating whether to print messages during processing. Default is TRUE.}

\item{include_latest}{A logical value indicating whether to include the latest time cutoff. Default is TRUE.}

\item{inclued_oldest}{A logical value indicating whether to include the oldest time cutoff. Default is FALSE.}

\item{input_tz}{A character value passed to lubridate::as_datetime indicating the input timezone. Default is "CEST".}
}
\value{
A named list of data.tables split by timeframe categories.
}
\description{
Prepares the post data to be past on to the comments-collector functions.
This function processes JSON files from a specified directory or file paths. It validates the JSON files,
parses the filenames to extract information, filters files based on a specified timeframe, and splits the
data into multiple data.tables based on timeframe categories.
}
