% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiktok_helpers.R
\name{slice_post_timeframes_tt}
\alias{slice_post_timeframes_tt}
\title{Read TikTok JSON post files and split data into timeframes}
\usage{
slice_post_timeframes_tt(
  dir = NULL,
  cutoff = NULL,
  recursive = FALSE,
  verbose = TRUE,
  include_latest = TRUE,
  include_oldest = FALSE,
  input_tz = "CEST"
)
}
\arguments{
\item{dir}{A character string specifying the directory from which data needs to be parsed.}

\item{cutoff}{A numeric vector of cutoff times in hours for categorizing the time difference to now.}

\item{recursive}{A logical value indicating whether to search the directory recursively. Default is \code{FALSE}.}

\item{verbose}{A logical value indicating whether to print messages during processing. Default is \code{TRUE}.}

\item{include_latest}{A logical value indicating whether to include the latest time cutoff. Default is \code{TRUE}.}

\item{include_oldest}{A logical value indicating whether to include the oldest time cutoff. Default is \code{FALSE}.}

\item{input_tz}{A character value passed to \code{lubridate::as_datetime} indicating the input timezone. Default is \code{"CEST"}.}
}
\value{
A named list of data.tables split by timeframe categories.
}
\description{
Prepares the TikTok data to be passed on to the comments-collector functions.
This function processes JSON files from a specified directory or file paths. It validates the JSON files,
parses the filenames to extract information, filters files based on a specified timeframe, and splits the
data into multiple data.tables based on timeframe categories.
}
\examples{
\dontrun{
dir <- "path/to/your/directory"
cutoff <- c(1, 24, 72, 240)
result <- slice_post_timeframes_tt(dir = dir, cutoff = cutoff, recursive = TRUE)
}
}
