% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged_calls.R
\name{call_log_ct}
\alias{call_log_ct}
\title{Call Crowdtangle API and Log Results}
\usage{
call_log_ct(
  grid_df,
  traceback = FALSE,
  autolog = TRUE,
  show_notes = TRUE,
  compact = FALSE,
  progress_bar = TRUE,
  return_results = TRUE,
  verbose = TRUE,
  fb_token,
  ig_token
)
}
\arguments{
\item{grid_df}{The grid dataframe containing information about the API calls to be made.}

\item{traceback}{Logical; indicates whether to show traceback in case of an error. Default is FALSE.}

\item{autolog}{Logical; indicates whether to start logging. Default is TRUE.}

\item{show_notes}{Logical; indicates whether to show informational notes during logging. Default is TRUE.}

\item{compact}{Logical; indicates whether to use a compact logging format. Default is FALSE.}

\item{progress_bar}{Logical; indicates whether to display a progress bar during API calls. Default is TRUE.}

\item{return_results}{Logical; indicates whether to return the results of the API calls. Default is TRUE.}

\item{verbose}{Logical; indicates whether to display verbose output during logging. Default is TRUE.}

\item{fb_token}{The Facebook API token.}

\item{ig_token}{The Instagram API token.}
}
\value{
If return_results is TRUE, a list containing the results of the API calls.
}
\description{
This function calls the Crowdtangle API for each row in the provided grid dataframe, logs the results, and optionally returns the results.
}
