% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_call_grid.R
\name{create_call_grid}
\alias{create_call_grid}
\title{Create Call Grid for Social Media Data Collection}
\usage{
create_call_grid(
  df = df,
  platform = c("fb", "ig", "tt", "yt", "tg", "bc", "bs"),
  country_var = NULL,
  party_var = NULL,
  name_var = NULL,
  filename_var = NULL,
  account_var = NULL,
  start_date = NULL,
  end_date = NULL,
  unit = "month",
  name_sep = "-",
  lowercase = TRUE,
  replace_non_ascii = TRUE,
  filename_sep = "_",
  sortBy = "date",
  parse = TRUE,
  data_path = NULL,
  count = Inf,
  drop_existing = FALSE
)
}
\arguments{
\item{df}{A data frame containing the relevant data.}

\item{platform}{A character vector specifying the platform(s) from which the data is collected.
Default is c("fb", "ig", "tt", "yt", "tg", "bc", "bs").}

\item{country_var}{A character string specifying the column name in the data frame that contains country information.
Default is NULL.}

\item{party_var}{A character string specifying the column name in the data frame that contains party information.
Default is NULL.}

\item{name_var}{A character string specifying the column name in the data frame that contains name information.
Default is NULL.}

\item{filename_var}{A character string specifying the column name in the data frame that contains filename information.
Default is NULL.}

\item{account_var}{A character string specifying the column name in the data frame that contains the account specifier - ID or handle.
Default is NULL.}

\item{start_date}{A character or Date object indicating the starting date of the timeframe.}

\item{end_date}{A character or Date object indicating the ending date of the timeframe.}

\item{unit}{A character vector specifying the unit of the time intervals.
Options include "day", "week", "month", "quarter", and "year". Default is "month".}

\item{name_sep}{A character string specifying the separator for name variables. Default is "-".}

\item{lowercase}{Logical; if TRUE, filenames will be converted to lowercase. Default is TRUE.}

\item{replace_non_ascii}{Logical; if TRUE, non-ASCII characters will be replaced in filenames. Default is TRUE.}

\item{filename_sep}{A character string specifying the separator for filename construction. Default is "_".}

\item{sortBy}{A character string specifying the parameter by which the data will be sorted.}

\item{parse}{Logical; if TRUE, the data will be parsed. Default is TRUE.}

\item{data_path}{A character string indicating the path where the data will be stored.}

\item{count}{An integer specifying the number of records to fetch. Default is Inf.}

\item{drop_existing}{Logical. Should calls for files that already exist in data_path be excluded from the grid. Default is FALSE.
Default is NULL, which means the data will be stored in the current directory under the "data" folder.}
}
\value{
A data frame containing a grid of account ids or handles and time frames for data collection.
}
\description{
This function constructs a grid of account handles or IDs and time frames to collect data from various APIs,
especially useful for social media platforms.
}
\examples{
df <- tibble::tribble(
  ~country, ~party, ~name, ~handle,
  "at", "fpö", "Herbert Kickl", "herbertkickl",
  "at", "övp", "TBD", NA,
  "at", "Grüne", "Werner Kogler", "wernerkogler"
)
create_call_grid(
  df = df,
  platform = "fb",
  country_var = "country",
  party_var = "party",
  name_var = "name",
  account_var = "handle",
  filename_var = "filename",
  start_date = "2023-08-25",
  end_date = "2024-01-24",
  unit = "quarter"
)

}
