% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cutoffs}
\alias{create_cutoffs}
\title{Create Breaks and Labels for Time Intervals}
\usage{
create_cutoffs(cutoff = NULL)
}
\arguments{
\item{cutoff}{A numeric vector of cutoff values for the intervals. If not provided, defaults to \code{NULL}.}
}
\value{
A list containing two elements:
\itemize{
\item \code{cutoff}: A numeric vector of cutoff values for the intervals, sorted in ascending order and containing \code{Inf}.
\item \code{labels}: A character vector of labels for the intervals.
}
}
\description{
This function constructs a vector of breaks and corresponding labels for specified time intervals.
}
\details{
If \code{cutoff} is \code{NULL} or empty, the function sets \code{cutoff} to \code{Inf}.
It then constructs the breaks vector starting from the first element of \code{cutoff} and
includes all elements of \code{cutoff}. The labels are constructed based on the intervals
defined by \code{cutoff}. If the start of the interval (\code{.x}) is less than 0 or equals
\code{Inf}, no dash "-" is included in the label.
}
