% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_call_grid.R
\name{drop_redundant}
\alias{drop_redundant}
\title{Drop Redundant Rows from a Data Frame}
\usage{
drop_redundant(df, account_var = NULL)
}
\arguments{
\item{df}{A data frame to be processed.}

\item{account_var}{A character string specifying the column name in the data frame based on which redundancy is checked. Default is NULL.}
}
\value{
A data frame with redundant rows removed based on the specified variable.
}
\description{
This function removes redundant rows from a data frame based on a specified variable.
}
\examples{
df <- data.frame(handle = c("herbertkickl", NA, "wernerkogler", "wernerkogler"))
drop_redundant(df, account_var = "handle")

}
