% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_stored_data.R
\name{find_latest}
\alias{find_latest}
\title{Find the latest file for each account in the specified path}
\usage{
find_latest(path)
}
\arguments{
\item{path}{A character string specifying the directory path where the files are located.}
}
\value{
A data.table with columns representing account details and the latest file information.
}
\description{
This function parses the latest files for each account found in the specified path.
It returns a data.table containing the latest file for each unique account.
}
