% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiktok_helpers.R
\name{get_comments_tt}
\alias{get_comments_tt}
\title{Call TikTok API to get comment metadata of the collected videos and Log Results}
\usage{
get_comments_tt(recent_videos, work_dir)
}
\arguments{
\item{recent_videos}{MM pls add doc.}

\item{work_dir}{Sets a working directory - MM plz check.}
}
\value{
If return_results is TRUE, a list containing the results of the API calls.
}
\description{
This function calls the TikTok API for each row in the provided grid dataframe, logs the results, and optionally returns the results.
}
