% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_stored_data.R
\name{parse_latest}
\alias{parse_latest}
\title{Parse Latest JSON Files}
\usage{
parse_latest(path)
}
\arguments{
\item{path}{A character string specifying the path where the JSON files are located.}
}
\value{
A data.table containing parsed information from the latest JSON files.
}
\description{
This function parses the latest JSON files based on the datetime information in the filenames.
}
