% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_error_jsons}
\alias{remove_error_jsons}
\title{Remove and Move Corrupted JSON Files}
\usage{
remove_error_jsons(dir, filepaths = NULL)
}
\arguments{
\item{dir}{A character string specifying the directory containing the JSON files to be checked.}

\item{filepaths}{A character string specifying the full file paths the JSON files. Default is NULL.}
}
\value{
This function does not return a value. It moves corrupted JSON files to a subdirectory
and prints a warning message if any corrupted files are found.
}
\description{
This function scans a specified directory for JSON files, checks their validity based on the \code{status} field,
and moves any corrupted files (with \code{status} not equal to 200) to a subdirectory named \code{corrupted_files}.
}
\details{
The function first verifies the existence of the specified directory. It then searches
for all JSON files in the directory and validates them based on their \code{status} field. JSON files
with a \code{status} field not equal to 200 are considered corrupted and moved to a \code{corrupted_files}
subdirectory within the specified directory.
}
