% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_invalid_jsons}
\alias{remove_invalid_jsons}
\title{Remove Invalid JSON Files}
\usage{
remove_invalid_jsons(dir)
}
\arguments{
\item{dir}{A character string specifying the directory containing the JSON files.}
}
\value{
This function does not return anything; it removes invalid JSON files by moving them to the "corrupted_files" sub-directory.
}
\description{
This function checks JSON files in a specified directory for validity. If any invalid JSON files are found, they are moved to a sub-directory named "corrupted_files" within the specified directory.
}
