% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_call_grid.R
\name{slice_timeframes}
\alias{slice_timeframes}
\title{Slice Timeframes}
\usage{
slice_timeframes(
  start_date = NULL,
  end_date = NULL,
  unit = c("day", "week", "month", "quarter", "year")
)
}
\arguments{
\item{start_date}{The starting date for the sequence of timeframes. Default is NULL.}

\item{end_date}{The ending date for the sequence of timeframes. Default is NULL.}

\item{unit}{The time unit for slicing the timeframes: "day", "week", "month", "quarter", "year". Default is "day".}
}
\value{
A data frame containing the sliced timeframes with columns:
\itemize{
\item \code{start_date}: The start date of each timeframe.
\item \code{start_datetime}: The start datetime of each timeframe in ISO 8601 format.
\item \code{end_date}: The end date of each timeframe.
\item \code{end_datetime}: The end datetime of each timeframe in ISO 8601 format.
}
}
\description{
Generate a sequence of timeframes based on specified start and end dates.
}
\details{
This function generates a sequence of timeframes based on the specified start and end dates.
If start_date or end_date is not provided, default values are used. If end_date is in the future,
it is replaced by the current date. If start_date is after end_date, it is adjusted accordingly.
}
\examples{
slice_timeframes(start_date = "2023-01-01", end_date = "2023-12-31", unit = "month")

}
