% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_call_grid.R
\name{slice_timeframes}
\alias{slice_timeframes}
\title{Slice Timeframes into Intervals}
\usage{
slice_timeframes(
  start_date = NULL,
  end_date = NULL,
  unit = c("day", "week", "month", "quarter", "year")
)
}
\arguments{
\item{start_date}{A character or Date object indicating the starting date of the timeframe.
If NULL, it defaults to 1 week before today.}

\item{end_date}{A character or Date object indicating the ending date of the timeframe.
If NULL, it defaults to today.}

\item{unit}{A character vector specifying the unit of the time intervals.
Options include "day", "week", "month", "quarter", and "year". Default is "day".}
}
\value{
A data frame containing sliced timeframes with start and end dates,
along with corresponding start and end datetimes.
}
\description{
This function slices timeframes into intervals specified by the user.
}
\examples{
slice_timeframes(start_date = "2023-01-01", end_date = "2023-12-31", unit = "month")

}
